Rework ActionUtility
Most methods of ActionUtility are build for menus but accept IAction.
Therefore, they do not handle child menus as IAction does not know child
actions. Move those methods to MenuUtility, change their signature to
IMenu and migrate all usages. As a consequence also the signature of
MenuWrapper, OutlineMenuWrapper and OutlineFormMenuWrapper are changed
accordingly.

240403